[snackager] Fix failing findPath test due to missing folder

+++ snackager/src/utils/__tests__/findPath.test.ts
@@ -11,7 +11,9 @@ const mkdirSync = (...segments): void => {
 };
 
 afterEach(() => {
-  fs.rmdirSync(tmpDir, { recursive: true });
+  if (fs.existsSync(tmpDir)) {
+    fs.rmdirSync(tmpDir, { recursive: true });
+  }
 });
 
 it('returns no paths for non existing package', () => {
